+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.8.0
+
+* Implements new cookie manager for setting cookies and providing initial cookies.
+
+## 2.7.0
+
+* Adds support for the `loadRequest` method from the platform interface.
+
 ## 2.6.0
 
 * Adds implementation of the `loadFlutterAsset` method from the platform interface.